requirements.txt
azure-functions
azure-identity
azure-mgmt-cosmosdb
pytz
